Add files via upload
With the advent of embedded systems Technology, we are at the doorstep of domesticating
Advanced Technological concepts incorporated into products that include (but are not limited to)
embedded systems, the internet of things, connected devices, computer networks ,mobile
communication etc. This statement is justified by the rising trend in home automation and connected
devices in the Western world ,the Eastern part of the globe seems to have been left behind in the
race owing to lack of options in the market, high costs and maintenance charges, need to change
in initial wiring bulky and monotonous systems sold in the name of Home Automation.
Our project aims at bringing Home Automation systems to the common household with absolute
ease of use. Our project is based on the following USP's (unique selling points) that are
Affordability, Versatility Retrofit and State of the art systems designed to provide seamless voice
control Feather Touch "plug and play" home automation system with the least possible costs and
fuss.
The idea here is to create a practical Home automation board that can fit into our AC power units
on our walls and stay concealed inside it. The board should not interrupt the normal working of our
power unit switches, that is they should turn ON or OFF with manual switches as well. And without
being said, it should also be able to control the same load with voice commands too. Since this
brings in various technologies, the project gets interesting and practical to use.
The system would also comprise of various other features such as voice control and will also have
sensors integrated into the board, which would complement the home automation system. The aim
is to build a prototype which would lead to a product solving a common need in home automation
in households with pre-existing wirings and installed switchboards.
Our project aims at bringing Home Automation systems to the common household with absolute
ease of use. Our project is based on the following USP's (unique selling points) that are
Affordability,Versatility Retrofit and State of the art systems designed to provide seamless voice
control Feather Touch "plug and play" home automation system with the least possible costs and
fuss.
The idea here is to create a practical Home automation board that can fit into our AC power unit
son our walls and stay concealed inside it. The board should not interrupt the normal working of
our power unit switches,that is they should turn ON or OFF with manual switches as well.And
without being said, it should also be able to control the same load with voice commands too.Since
this brings in various technologies,the project gets interesting and
Practical to use.
The system would also comprise of various other features such as voice control and will also
have sensors integrated into the board, which would complement the home automation system.
Thus. by using technologies like IOT, cloud computing and mobile commu nications and
Embedded systems we can make Home automation experience of customers smooth and easy as
well as make empower other sectors where this systems can be incorporated .